% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/add_crossSections_to_reach.R
\name{add_crossSections_to_reach}
\alias{add_crossSections_to_reach}
\title{Add cross-sections to a hecras reach with a shapefile and a DEM}
\usage{
add_crossSections_to_reach(hec_geo, xsect_lines, dem, dem_vertical_offset = 0,
  reaches_to_modify_pattern = NULL, outfile = NULL, verbose = FALSE,
  overwrite = FALSE, station_name_offset = 0, default_manning = 0.03, ...)
}
\arguments{
\item{hec_geo}{Either a filename (pointing to a hecras geometry file), or a character vector containing the hecras geometry. The latter could come from 'readLines(hecras_geometry_file)'}

\item{xsect_lines}{Either a line shapefile name (containing the plan-view geometries of the cross-sections we want to add to the reach), or a SpatialLines* object containing those geometries}

\item{dem}{Either a raster filename corresponding to the DEM, or a rasterLayer}

\item{dem_vertical_offset}{A number to add to the DEM values during processing}

\item{reaches_to_modify_pattern}{Character vector. Reaches will only be modified if they grep at least one pattern in reaches_to_modify_pattern}

\item{outfile}{Filename to save the final geometry to}

\item{verbose}{TRUE/FALSE print progress messages}

\item{overwrite}{TRUE/FALSE -- overwrite outfile if it already exists?? Be careful}

\item{station_name_offset}{-- constant to add to the station_names (which will be real numbers). Useful because when merging channels, hecras requires the more usptream stations to have numbers greater than the more downstream stations.}

\item{default_manning}{Manning's n to use on the cross-section. Only a single value is supported, but see edit_manning_from_shapefile}

\item{...}{Any other arguments to 'compute_crossSection_elev_with_dem'}
}
\value{
If outfile is NULL, then return a character vector containing the updated hecras geometry. Otherwise, write the geometry to a file named 'outfile'.
}
\description{
The reach must already have a centreline in the hecras geometry,
and the reaches which will be modified should not contain any cross-sections. Not all important
inputs can be specified (junctions / manning / ect) so some further editing will be required.
}
\details{
WARNING: The cross-sectional profiles are computed with 'compute_crossSection_elev_with_dem'. Only a single downstream distance is set, and manning's n values are not specified. Use this to quicky get data into hec-ras, but assume you still need to check/edit the inputs.
}
\examples{
partial_hecras_geofile=paste0(system.file(package='RgeoRAS'), '/extdata/centreline_only/make_xsect.g01')
  xsect_lines=paste0(system.file(package='RgeoRAS'), '/extdata/centreline_only/xsects/xsects.shp')
  lidar_DEM_file=paste0(system.file(package='RgeoRAS'), '/extdata/10m_dem.tif')
  lidar_DEM=raster(lidar_DEM_file)
  # Add cross-sections -- note use of optional argument to 'compute_crossSection_elev_with_dem' to control spacing between cross-sectional points
  new_ras_file=add_crossSections_to_reach(partial_hecras_geofile, xsect_lines,
                   lidar_DEM, reaches_to_modify_pattern='*', desired_spacing=3.0)
  # print(new_ras_file)
  # cat(new_ras_file,'output.g01', sep="\\r\\n") # Output with carriage return good for windows
}

