% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/compute_hec_util.R
\name{check_storage_connectedness}
\alias{check_storage_connectedness}
\title{Check that storage areas are connected to something}
\usage{
check_storage_connectedness(hecras_file)
}
\arguments{
\item{hecras_file}{filename of hecras geometry}
}
\value{
A print statement if all storage areas are connected. Otherwise,
        browser() is called so you can check what happens
}
\description{
In hec-ras, weird things can happen if you have a storage area with no hydraulic connections
This function checks that all storage areas are connected to something
}

