% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/compute_hec_util.R
\name{compute_storage_stage_vol_relation}
\alias{compute_storage_stage_vol_relation}
\title{Compute a stage-volume relation for a polygon from a DEM}
\usage{
compute_storage_stage_vol_relation(my_poly, lidar_DEM,
  vertical_datum_offset = 0, upper_bound_stage_offset = 100)
}
\arguments{
\item{my_poly}{A SpatialPolygons or SpatialPolygonsDataFrame object with length(my_poly)==1}

\item{lidar_DEM,}{A rasterLayer digital elevation model in a projected coordinate system. res(lidar_DEM) is in units of metres}

\item{vertical_datum_offset}{A constant that is added to the lidar_DEM values before output}

\item{upper_bound_stage_offset}{The maximal stage in the returned stage-volume
  curve will be max(elevation_in_mypoly)+upper_bound_stage_offset. Set it to ensure the highest possible water surface
  elevation in the storage area is not exceeded}
}
\value{
A 2-column matrix with a stage-Volume relation. Compared with the DEM, with the stage values offset by + vertical_datum_offset
}
\description{
Function to compute the stage-volume relation for a region defined by the polygon my_poly
 , on a given lidar_Dem raster
}
\examples{
require(raster)
    require(rgdal)
    mychanfile=paste0(system.file(package='RgeoRAS'), '/extdata/marikina.g36')
    mydem=paste0(system.file(package='RgeoRAS'), '/extdata/10m_dem.tif')
    lidar_DEM=raster(mydem)
    myStoragePoly=spatial_storage_areas(mychanfile, CRS('+init=epsg:3123'))
    myStageVolRel=compute_storage_stage_vol_relation(myStoragePoly[1,], lidar_DEM)
    plot(myStageVolRel,xlab='Stage',ylab='Volume')
}

