% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/edit_manning_from_shapefile.R
\name{edit_manning_from_shapefile}
\alias{edit_manning_from_shapefile}
\title{Code to set manning's n values from shapefile(s)}
\usage{
edit_manning_from_shapefile(hec_geo, SpPolys_list, mann_list,
  preference_order = 1:length(mann_list), reaches_to_modify_pattern = "*",
  output_file = NULL, verbose = TRUE)
}
\arguments{
\item{hec_geo}{Filename of hecras geometry, or a character vector produced from such a file}

\item{SpPolys_list}{Either a list of shapefile filenames, OR a list of SpatialPolygons* objects, OR a single SpatialPolygons* object}

\item{mann_list}{A list or vector of manning's n values inside polygons in SpPolys_list. Must have the same length as SpPolys_list, with e.g. mann_list[[1]] being the manning n value inside SpPolys[[1]]}

\item{preference_order}{Vector of unique integers giving the 'preference-rank' for each SpPoly, which is used to set manning's n in case of overlaps. The length(preference_order) = length(SpPolys_list), and e.g. preference_order[1] is the rank of SpPolys_list[[1]], preference_order[2] is the rank of SpPolys_list[[2], etc. Polygons with lower values in preference_order are preferred to those with higher values.}

\item{reaches_to_modify_pattern}{Vector of one or more characters. Only reaches with names that pattern-match one or more elements of reaches_to_modify_pattern will be affected by the function.}

\item{output_file}{File to write the geometry to. If NULL, return the modified geometry as a character vector}

\item{verbose}{Print many messages}
}
\value{
The modified geometry as a character vector, unless output_file!=NULL, in which case the modified geometry is written to a output_file.
}
\description{
Code to set manning's n values from shapefile(s)
}
\details{
Pattern-matching using \code{reaches_to_modify_pattern} is done by grep. Defaults to '*', which matches everything. For example, say we have three reaches 'Reach1', 'NewRiver', 'Reach2'. Then reaches_to_modify_pattern=c('Reach', 'New') will modify everything, reaches_to_modify_pattern=c('2') will only modify 'Reach2', reaches_to_modify_pattern='X' will not modify anything, etc.
}
\examples{
mychanfile=paste0(system.file(package='RgeoRAS'), '/extdata/Geo_with_xsect.g02')
    manning_0p03=paste0(system.file(package='RgeoRAS'), '/extdata/Manning/Lower_delta.shp')
    manning_0p05=paste0(system.file(package='RgeoRAS'), '/extdata/Manning/Mann_chan.shp')
    manning_0p1=paste0(system.file(package='RgeoRAS'), '/extdata/Manning/Mann1.shp')
    # Prefer to use 0p03, then 0p05, then 0p1
    new_geo=edit_manning_from_shapefile(mychanfile,
                                        list(manning_0p05, manning_0p03, manning_0p1),
                                        list(0.05, 0.03, 0.1),
                                        preference_order=c(2, 1, 3))
}

