% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/text_util.R
\name{format_in_rows}
\alias{format_in_rows}
\title{Format in rows}
\usage{
format_in_rows(char_vec, rowlen)
}
\arguments{
\item{char_vec}{A character vector}

\item{rowlen}{The number of entries in each row}
}
\value{
A character vector
}
\description{
Function to re-format a character vector to have rows containing 'rowlen' entries
e.g for rowlen=2
c(' 01', ' 23', ' 34.5', 'my char', 'f') --> c(' 01 23', ' 34.5my char', 'f')
}
\examples{
x=c(' 01', ' 23', ' 34.5', 'my char', 'f')
  y=format_in_rows(x, 2)
  stopifnot(y==c(' 01 23', ' 34.5my char', 'f'))
}

