% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/text_util.R
\name{heclines_2_File}
\alias{heclines_2_File}
\title{Write character vector containing hecras geometry to a file, in a way that works from both linux and windows machines.}
\usage{
heclines_2_File(hec_lines, output_file)
}
\arguments{
\item{hec_lines}{character vector containing hecras geometry info}

\item{output_file}{output filename}
}
\value{
Nothing, but saves the file
}
\description{
Write character vector containing hecras geometry to a file, in a way that works from both linux and windows machines.
}

