% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/make_text_hec_util.R
\name{make_crossSection_text}
\alias{make_crossSection_text}
\title{Make hecras geometry text for a cross-section}
\usage{
make_crossSection_text(station_name, section_cutline, section,
  downstream_distance, mann = 0.03, HTAB_increm = 0.2)
}
\arguments{
\item{station_name}{-- the name of the station (at most 8 characters)}

\item{section_cutline}{-- 2-column matrix with x-y coordinates defing the spatial locations of the cross-section}

\item{section}{-- 2-column matrix with distance-elevation coordinates defining the cross-section profile}

\item{downstream_distance}{-- the downstream distance of the cross-section. Can be a vector of length 3, with values for the left bank,channel,right bank. Will be rounded to 2 decimal places.}

\item{mann}{-- (not implemented) manning's n value}

\item{HTAB_increm}{-- increment for the HTAB parameters}
}
\value{
text defining the cross-section that can be inserted in a hecras geometry file
}
\description{
Make hecras geometry text for a cross-section
}
\examples{
require(raster)
 # Read base data
 mydem=paste0(system.file(package='RgeoRAS'), '/extdata/10m_dem.tif')
 lidar_DEM=raster(mydem)

 # Make an artificial section cutline  (would typically use real data with many more points)
 section_cutline=matrix(c(509619.4, 511972.1, 1610852, 1613042), ncol=2)
 len_section_cutline= sqrt(diff(section_cutline[,1])**2 + diff(section_cutline[,2])**2)
 # Make an artificual section profile along section cutline
 section=matrix(c(seq(0,len_section_cutline,len=4), seq(-20,-20,len=4)),ncol=2)

 # Apply the function
 newsection=compute_crossSection_elev_with_dem(section,section_cutline, lidar_DEM)
 outtext=make_crossSection_text('100.102', section_cutline,
                    newsection, downstream_distance=c(195,200.0,197))
 print(outtext)
}

