% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/make_text_hec_util.R
\name{make_lateral_weir_text}
\alias{make_lateral_weir_text}
\title{Make lateral weir text from storage area / channel gis information}
\usage{
make_lateral_weir_text(storage_poly, storage_name, chan_poly, chan_pts,
  lidar_DEM, vertical_datum_offset = 0, hec_lines, bridge_lines,
  limit_weir_elevation_by_channel_bank_elevation, min_structure_elev,
  lower_limit_on_lateral_weir_elevations)
}
\arguments{
\item{storage_poly}{storage area polygon of length 1}

\item{storage_name}{name of storage area}

\item{chan_poly}{channel polygon from spatial_channel_polygon}

\item{chan_pts}{channel boundary points from spatial_channel_boundary_points}

\item{lidar_DEM}{Digital elevation model raster}

\item{vertical_datum_offset}{Number to add to lidar_DEM elevations}

\item{hec_lines}{vector of character strings containing hecras geometry file}

\item{bridge_lines}{lines in hec_lines matching "Bridge Culvert"}

\item{limit_weir_elevation_by_channel_bank_elevation}{TRUE/FALSE -- place a (local) lower limit on a weir elevation}

\item{min_structure_elev}{[NOT IMPLEMENTED] another limit on the storage elevation}

\item{lower_limit_on_lateral_weir_elevations}{A global lower limit on the lateral weir elevations}
}
\value{
modified version of hec_lines
}
\description{
For adding lateral weirs to hec-ras files, when storage areas are added with polygons
}

