% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/make_text_hec_util.R
\name{make_storage_area_text}
\alias{make_storage_area_text}
\title{Make text representing a storage area in hec-ras format}
\usage{
make_storage_area_text(storage_area, elev_vol, name)
}
\arguments{
\item{storage_area}{-- A SpatialPolygons/SpatialPolygonsDataFrame representing a single storage area}

\item{elev_vol}{-- A stage-volume relation (typically output from 'compute_storage_stage_vol_relation')}

\item{name}{-- A name for the storage area that hec-ras will use}
}
\value{
A lot of text
}
\description{
Function to make a block of text representing a storage area,
as appears in the hec-ras format .g01
Probably, this will only be of internal use
}
\examples{
require(raster)
    require(rgdal)
    require(rgeos)
    mychanfile=paste0(system.file(package='RgeoRAS'), '/extdata/marikina.g36')
    mydem=paste0(system.file(package='RgeoRAS'), '/extdata/10m_dem.tif')
    lidar_DEM=raster(mydem)
    myStoragePoly=spatial_storage_areas(mychanfile, CRS('+init=epsg:3123'))
    myStageVolRel=compute_storage_stage_vol_relation(myStoragePoly[1,], lidar_DEM)
    #sat=make_storage_area_text(myStoragePoly[1,], myStageVolRel, 'Blah')
    #print(sat)
}

