% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/make_text_hec_util.R
\name{make_storage_connection_text}
\alias{make_storage_connection_text}
\title{Make hecras text for a storage connection}
\usage{
make_storage_connection_text(store1, store2, name1, name2, lidar_DEM,
  vertical_datum_offset = 0,
  storage_area_overlap_merge_threshold = glob_vars$storage_area_overlap_merge_threshold)
}
\arguments{
\item{store1}{SpatialPolygonsDataFrame of length(1) -- the first storage area to join}

\item{store2}{The second storage area to join}

\item{name1}{The name of the first storage area}

\item{name2}{The name of the second storage area}

\item{lidar_DEM}{A raster digital elevation model, projected coordinate system, with res(lidar_DEM) in metres}

\item{vertical_datum_offset}{A constant to be added to the lidar_DEM values before processing}

\item{storage_area_overlap_merge_threshold}{If the storage areas overlap by > this fraction of the area of the smaller storage area, throw an error.}
}
\value{
lots of text
}
\description{
Make hecras text for a storage area connection, with lots of assumed parameters (careful).
Typically used to get a connection into the geometry, but then the parameter
values must be edited separately
}
\examples{
require(raster)
  require(rgdal)
  require(rgeos)
  mychanfile=paste0(system.file(package='RgeoRAS'), '/extdata/marikina.g36')
  mydem=paste0(system.file(package='RgeoRAS'), '/extdata/10m_dem.tif')
  lidar_DEM=raster(mydem)
  myStoragePoly=spatial_storage_areas(mychanfile, CRS('+init=epsg:3123'))
  P1=myStoragePoly[1,]
  P2=myStoragePoly[536,]
  msct = make_storage_connection_text(P1, P2, 'Blah', 'Bloo', lidar_DEM, vertical_datum_offset=0)
  print(msct)
}

