% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/text_util.R
\name{pad_string}
\alias{pad_string}
\title{Format character strings to have a fixed length}
\usage{
pad_string(mystring_vec, charlen, pad = " ", justify = "left")
}
\arguments{
\item{mystring_vec}{character string vector}

\item{charlen}{Integer length to pad the string to}

\item{pad}{Character to pad the string with. Must be of length 1}

\item{justify}{'left' or 'right' justify the string}
}
\value{
A new character vector which is appropriately padded
}
\description{
Take a string, check its length, and if it is too short, then pad it to
the correct length. If it is too long, then truncate it.
}
\examples{
x=pad_string('Mystring', 16, '-', 'left')
 print(x)
 stopifnot(x=='Mystring--------')
 x=pad_string('Mystring', 16, ' ', 'right')
 print(x)
 stopifnot(x=='        Mystring')
}

