% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RgeoRAS-package.R
\docType{package}
\name{RgeoRAS-package}
\alias{RgeoRAS}
\alias{RgeoRAS-package}
\title{The EXPERIMENTAL RgeoRAS Package}
\description{
EXPERIMENTAL Tools to integrate HecRAS geometry files and GIS datasets, plus other useful things..
}
\details{
\tabular{ll}{
Package: \tab RgeoRAS\cr
Type: \tab Package\cr
Version: \tab 0.00-0\cr
Date: \tab 2013-11-08\cr
Depends: \tab R (>= 2.15.1), raster, rgdal, rgeos, sp\cr
Encoding: \tab UTF-8\cr
License: \tab BSD3 \cr
LazyLoad: \tab yes\cr
URL: \tab NONE\cr
}

Functions starting with 'spatial_' create spatial geometries from hecras geometry files.
Functions starting with 'make_' create hecras geometry text which can be inserted into hecras geometry files.
Functions starting with 'add_, edit_, update_' etc modify a character vector representing a hecras geometry file (produced from readLines(hecras_geometry_file)).
Functions starting with 'SCS_' relate to the SCS unit hydrograph rainfall runoff model.
Other functions exist too.
}
\author{
Gareth Davies \email{gareth.davies.ga.code@gmail.com}
}
\keyword{package}

