% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scs_model.R
\name{SCS_runoff_volume}
\alias{SCS_runoff_volume}
\title{Compute runoff volume using the SCS runoff model}
\usage{
SCS_runoff_volume(time_increment, precip_depth, CN = 74, Area = 1,
  dt = time_increment, initial_abstraction_coef = 0.2)
}
\arguments{
\item{time_increment}{time over which precip_depth is measured (s)}

\item{precip_depth}{vector of 'depth of rainfall' (mm)}

\item{CN}{curve number}

\item{Area}{catchment area (km^2)}

\item{dt}{time step over which to interpolate the data before starting. dt is used to be compatible with hec-hms}

\item{initial_abstraction_coef}{Coefficient for initial abstraction}
}
\value{
2-column matrix with 'time', 'incremental runoff volume' (in m^3 per dT -- like HMS)
}
\description{
See page 37+, hec-hms manual. For a comparison with hec-hms, see test_SCS_runoff()
}

