% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scs_model.R
\name{SCS_uh}
\alias{SCS_uh}
\title{Filter incremental rainfall excess through SCS unit hydrograph.}
\usage{
SCS_uh(diff_Pe, t_lag = 3600)
}
\arguments{
\item{diff_Pe}{time & Incremental rainfall excess (s, m^3) [output of runoff_volume_SCS]}

\item{t_lag}{lag for unit hydrograph (in seconds)}
}
\value{
time-series with 'time' (s), 'discharge' (m^3/s)
}
\description{
Dimensionless unit hydrograph is based on data from
http://www.nohrsc.noaa.gov/technology/gis/uhg_manual.html
For a comparison with hec-hms, see test_SCS_uh()
}

