% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spatial_hec_util.R
\name{spatial_channel_boundary_points}
\alias{spatial_channel_boundary_points}
\title{Make a shapefile of channel boundary points from a hecras file}
\usage{
spatial_channel_boundary_points(hec_chan_file, spatial_proj)
}
\arguments{
\item{hec_chan_file}{File name of hecras geometry, or a vector containing the contents of a file}

\item{spatial_proj}{CRS object defining the spatial coordinate system, OR a
character string with the proj4string info (in which case CRS(spatial_proj) is used )}
}
\value{
SpatialPointsDataFrame with boundary points of xsections
       The attribute table includes "reach_name", "station_name", "bank",
        "downstream_distance", "bank_elev", "line_num" (a cross-section id)
}
\description{
Make a shapefile of channel boundary points from a hecras file. The boundary
points define the end-points of the x-sections
}
\examples{
require(rgdal)
  mychanfile=paste0(system.file(package='RgeoRAS'), '/extdata/marikina.g36')
  my_boundary_points=spatial_channel_boundary_points(mychanfile, CRS("+init=epsg:3123"))
}

