% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spatial_hec_util.R
\name{spatial_channel_cutlines}
\alias{spatial_channel_cutlines}
\title{Make a shapefile of channel cross-sections in plan view (cutlines) from a hecras file}
\usage{
spatial_channel_cutlines(hec_chan_file, spatial_proj)
}
\arguments{
\item{hec_chan_file}{File name of hecras geometry, or the content of such a file in a character vector}

\item{spatial_proj}{CRS object defining the spatial coordinate system, OR a
character string with the proj4string info (in which case CRS(spatial_proj) is used )}
}
\value{
SpatialPointsDataFrame with boundary points of xsections
       contains variables "id_2","sec_info","reach", with info from the hecras file
}
\description{
Make a shapefile of channel cross-sections from a hecras file.
}
\examples{
require(rgdal)
  mychanfile=paste0(system.file(package='RgeoRAS'), '/extdata/marikina.g36')
  my_cutlines=spatial_channel_cutlines(mychanfile, CRS("+init=epsg:3123"))
}

