% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spatial_hec_util.R
\name{spatial_crossSection_points}
\alias{spatial_crossSection_points}
\title{Make a shapefile of channel cross-section points from a hecras file}
\usage{
spatial_crossSection_points(hec_chan_file, spatial_proj)
}
\arguments{
\item{hec_chan_file}{File name of hecras geometry, or a vector containing the contents of a file}

\item{spatial_proj}{CRS object defining the spatial coordinate system, OR a
character string with the proj4string info (in which case CRS(spatial_proj) is used )}
}
\value{
SpatialPointsDataFrame with points on xsections
       The attribute table includes "elevation", "station_name",
        "station_index", "reach_name"
}
\description{
Make a shapefile of points on channel cross-sections from a hecras file.
}
\examples{
require(rgdal)
  mychanfile=paste0(system.file(package='RgeoRAS'), '/extdata/marikina.g36')
  my_xsect_points=spatial_crossSection_points(mychanfile, CRS("+init=epsg:3123"))
}

