% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spatial_hec_util.R
\name{spatial_hecgeo2shp}
\alias{spatial_hecgeo2shp}
\title{Make shapefiles of key hecras geometries}
\usage{
spatial_hecgeo2shp(hec_chan_file, spatial_proj, outdir = NULL,
  overwrite = FALSE)
}
\arguments{
\item{hec_chan_file}{filename of hecras geometry, or a character vector containing the contents of such a file}

\item{spatial_proj}{either a CRS object containing the projection, or a character string which will be coerced to that with CRS( )}

\item{outdir}{directory to write the shapefiles to. By default this will be based on the name of hec_chan_file}

\item{overwrite}{TRUE/FALSE Overwrite existing shapefiles in outdir?}
}
\value{
None, but creates many shapefiles as a side-effect
}
\description{
Convenience function to make lots of shapefiles from hecras geometry
}
\examples{
\dontrun{
  mychanfile=paste0(system.file(package='RgeoRAS'), '/extdata/marikina.g36')
  myheclines=readLines(mychanfile)
  spatial_hecgeo2shp(myheclines,CRS("init=epsg:3123"))
}
}

