% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spatial_hec_util.R
\name{spatial_junction_polygon}
\alias{spatial_junction_polygon}
\title{Make a polygon of channel junctions}
\usage{
spatial_junction_polygon(hec_chan_file, channel_cutlines)
}
\arguments{
\item{hec_chan_file}{A hecras geometry file, or the contents of such a file as a character vector}

\item{channel_cutlines}{The output of spatial_channel_cutlines}
}
\value{
A SpatialPolygonsDataFrame with the junction geometries
       Contains variables "junct_name",  "xsect_inds",  "xsect_up_dn"
}
\description{
Make polygon of the junctions. This is needed for RAS Mapper type
plotting, and can be useful at other times too.
}
\examples{
require(rgdal)
  mychanfile=paste0(system.file(package='RgeoRAS'), '/extdata/marikina.g36')
  my_cutlines=spatial_channel_cutlines(mychanfile, CRS("+init=epsg:3123"))
  my_junctions=spatial_junction_polygon(mychanfile, my_cutlines)
}

