% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spatial_hec_util.R
\name{spatial_storage_areas}
\alias{spatial_storage_areas}
\title{Extract storage areas from a hec-ras file}
\usage{
spatial_storage_areas(hec_file, spatial_proj)
}
\arguments{
\item{hec_file}{Name of a hecras geometry file (typically with a .gXX
extension, where XX is an integer), or a character vector containing the
contents of such a file}

\item{spatial_proj}{A CRS object describing the projection (or a character
string, in which case CRS(spatial_proj) will be used)}
}
\value{
A SpatialPolygonDataFrame, with the storage area geometry, and its name in hec-ras
}
\description{
Function to extract storage areas from an existing hec-ras file
into a SpatialPolygonsDataFrame
Useful to plot them up, to help with the creation of new storage areas.
}
\examples{
require(rgdal)
  mychanfile=paste0(system.file(package='RgeoRAS'), '/extdata/marikina.g36')
  myStoragePoly=spatial_storage_areas(mychanfile, CRS('+init=epsg:3123'))
  plot(myStoragePoly,axes=TRUE)
}

