% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scs_model.R
\name{test_SCS_runoff}
\alias{test_SCS_runoff}
\title{Test the SCS runoff model by comparison with hec-hms outputs}
\usage{
test_SCS_runoff()
}
\value{
Some plots + a pass/fail statement
}
\description{
Test the SCS runoff model by comparison with hec-hms outputs
}

