% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scs_model.R
\name{test_SCS_uh}
\alias{test_SCS_uh}
\title{Test the SCS unit hydrograph function by comparison with hec-hms}
\usage{
test_SCS_uh()
}
\value{
Produce some comparison plots
}
\description{
Test the SCS unit hydrograph function by comparison with hec-hms
}

