% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/update_downstream_distances.R
\name{update_downstream_distances}
\alias{update_downstream_distances}
\title{Compute the 'downstream distances' of the xsections on each
reach, using the cutline and reach centreline information in hec_lines}
\usage{
update_downstream_distances(hec_lines, chan_cutlines, reach_lines)
}
\arguments{
\item{hec_lines}{vector of character strings from the hecras file (which may have been modified already)}

\item{chan_cutlines}{channel cutlines SpatialLinesDataFrame -- output from spatial_channel_cutlines}

\item{reach_lines}{reach centrelines SpatialLinesDataFrame -- output from spatial_centrelines}
}
\value{
modified version of hec_lines (with updated downstream distances)
}
\description{
Assumes that left bank distance = right bank distance = straight line
distance between x-sections at intersection with the channel,
while channel distance = distance along the channel.
}
\details{
NOTE: We are presently setting the left bank and right bank distances to be the same.
This should be appropriate if the overbank flows are largely occurring near the channel
However, this will not always be correct, although it will often be true
if the floodplain roughness gets high away from the channel.
In reality, the left bank / right bank distances should depend on the
flow state -- during serious inundation, they might differ quite a bit.
}
\examples{
#Nice_example
  x=2
  \dontrun{
      print('FIXME: The following fails')
      require(rgdal)
      mychanfile=paste0(system.file(package='RgeoRAS'), '/extdata/marikina.g36')
      hec_lines=readLines(mychanfile)
      my_reach_lines=spatial_centrelines(mychanfile, CRS("+init=epsg:3123"))
      my_cutlines=spatial_channel_cutlines(mychanfile,CRS("+init=epsg:3123"))
      usdists=update_downstream_distances(hec_lines, my_cutlines, my_reach_lines)
  }
}

