% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/compute_hec_util.R
\name{usdistfun}
\alias{usdistfun}
\title{Compute the upstream distance of a point on a line}
\usage{
usdistfun(point_coords, line_coords, roundoff_tol = 0.001)
}
\arguments{
\item{point_coords}{vector of length 2 with point coordinates}

\item{line_coords}{matrix with 2 columns defining the line. ASSUMES THAT FIRST POINT IS MOST UPSTREAM}

\item{roundoff_tol}{= Tolerence used to check if point is on the line}
}
\value{
The upstream distance
}
\description{
Compute the upstream distance of point='point_coords'=c(x0,y0), along a line
defined by 'line_coords' = matrix with 2 columns
Also, compute the index of the segment in line_coords that point_coords
intersects with. This was most convenient to calculate here, even though
its a bit out of place
point_coords should lie on line_coords to within a small tolerence. This
tolerance is is related to roundoff_tol in a complex fashion -- read the
source if you need to know!
NOTE: line_coords is assumed to be ordered from upstream to downstream.
Upstream distance is measured from the most downstream point in
line_coords
}

