% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/text_util.R
\name{vector_grep}
\alias{vector_grep}
\title{Crude extension of 'grep' to take multiple patterns.}
\usage{
vector_grep(patterns, text)
}
\arguments{
\item{patterns}{A character vector of patterns}

\item{text}{the text to search}
}
\value{
number of patterns that match the text. Note that this is cruder
         than 'grep', which returns the matching indices
}
\description{
Count how many patterns in 'patterns' matches the text in 'text'
}
\examples{
# Use of vector grep
  mypatterns=c('Hi', 'J', 'Elephant')
  mytext="Hi there, I would like to meet your Elephant"
  print(vector_grep(mypatterns,mytext))
}

